+++ Common.Core/ProceduralNoise/ValueNoise.cs
@@ -14,13 +14,13 @@ namespace Common.Core.ProceduralNoise
 	public class ValueNoise : Noise
 	{
 
-        public ValueNoise(int seed, double frequency, double amplitude = 1.0) 
+        public ValueNoise(int seed, float frequency, float amplitude = 1.0f) 
             : base(seed, frequency, amplitude)
         {
 
         }
 
-        public ValueNoise(int seed, Vector3d frequency, double amplitude = 1.0)
+        public ValueNoise(int seed, Vector3f frequency, float amplitude = 1.0f)
             : base(seed, frequency, amplitude)
         {
 
@@ -35,13 +35,13 @@ namespace Common.Core.ProceduralNoise
             return string.Format("[ValueNoise: Frequency={0}, Amplitude={1}, Offset={2}]", Frequency, Amplitude, Offset);
         }
 
-        public override double Sample1D(double x)
+        public override float Sample1D(float x)
         {
             x = (x + Offset.x) * Frequency.x;
 
             int ix0;
-            double fx0;
-            double s, n0, n1;
+            float fx0;
+            float s, n0, n1;
 
             ix0 = (int)Math.Floor(x);     // Integer part of x
             fx0 = x - ix0;                // Fractional part of x
@@ -52,19 +52,19 @@ namespace Common.Core.ProceduralNoise
             n1 = Perm[ix0 + 1];
 
             // rescale from 0 to 255 to -1 to 1.
-            double n = LERP(s, n0, n1) * Perm.Inverse;
-            n = n * 2.0 - 1.0;
+            float n = LERP(s, n0, n1) * Perm.Inverse;
+            n = n * 2.0f - 1.0f;
 
             return n * Amplitude;
         }
 
-        public override double Sample2D(double x, double y)
+        public override float Sample2D(float x, float y)
         {
             x = (x + Offset.x) * Frequency.x;
             y = (y + Offset.y) * Frequency.y;
 
             int ix0, iy0;
-            double fx0, fy0, s, t, nx0, nx1, n0, n1;
+            float fx0, fy0, s, t, nx0, nx1, n0, n1;
 
             ix0 = (int)Math.Floor(x);   // Integer part of x
             iy0 = (int)Math.Floor(y);   // Integer part of y
@@ -86,13 +86,13 @@ namespace Common.Core.ProceduralNoise
             n1 = LERP(t, nx0, nx1);
 
             // rescale from 0 to 255 to -1 to 1.
-            double n = LERP(s, n0, n1) * Perm.Inverse;
-            n = n * 2.0 - 1.0;
+            float n = LERP(s, n0, n1) * Perm.Inverse;
+            n = n * 2.0f - 1.0f;
 
             return n * Amplitude;
         }
 
-        public override double Sample3D(double x, double y, double z)
+        public override float Sample3D(float x, float y, float z)
         {
 
             x = (x + Offset.x) * Frequency.x;
@@ -100,9 +100,9 @@ namespace Common.Core.ProceduralNoise
             z = (z + Offset.z) * Frequency.z;
 
             int ix0, iy0, iz0;
-            double fx0, fy0, fz0;
-            double s, t, r;
-            double nxy0, nxy1, nx0, nx1, n0, n1;
+            float fx0, fy0, fz0;
+            float s, t, r;
+            float nxy0, nxy1, nx0, nx1, n0, n1;
 
             ix0 = (int)Math.Floor(x);   // Integer part of x
             iy0 = (int)Math.Floor(y);   // Integer part of y
@@ -136,15 +136,15 @@ namespace Common.Core.ProceduralNoise
             n1 = LERP(t, nx0, nx1);
 
             // rescale from 0 to 255 to -1 to 1.
-            double n = LERP(s, n0, n1) * Perm.Inverse;
-            n = n * 2.0 - 1.0;
+            float n = LERP(s, n0, n1) * Perm.Inverse;
+            n = n * 2.0f - 1.0f;
 
             return n * Amplitude;
         }
 
-        private double FADE(double t) { return t * t * t * (t * (t * 6.0 - 15.0) + 10.0); }
+        private float FADE(float t) { return t * t * t * (t * (t * 6.0f - 15.0f) + 10.0f); }
 
-        private double LERP(double t, double a, double b) { return a + t * (b - a); }
+        private float LERP(float t, float a, float b) { return a + t * (b - a); }
 
 	}
